fixed models, schemas, server, and all pages, components and src to have no errors, now need to fix it to render
