+++ client/src/App.jsx
@@ -1,98 +1,102 @@
-// // import './App.css';
-// // import { Outlet } from 'react-router-dom';
-// // import { ApolloClient, ApolloProvider, InMemoryCache } from '@apollo/client';
-
-// // const client = new ApolloClient({
-// //   uri: '/graphql',
-// //   cache: new InMemoryCache(),
-// // });
-
-// // function App() {
-// //   return (
-// //     <ApolloProvider client={client}>
-// //       <div className="flex-column justify-center align-center min-100-vh bg-primary">
-// //         <Outlet />
-// //       </div>
-// //     </ApolloProvider>
-// //   );
-// // }
-
-// // export default App;
-
 // import React from 'react';
-// import { BrowserRouter as Router, Switch, Route, Redirect } from 'react-router-dom';
-// import Login from './components/LoginForm';
-// import Signup from './components/SignupForm';
-// // import Home from './Home';
-// // import Profile from './Profile'; will fix these when my files are redone
-// function App() {
-//   const isLoggedIn = true;
+// import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
+// import { ApolloProvider, ApolloClient, InMemoryCache } from '@apollo/client';
+// import Header from './components/Header/Header';
+// import Footer from './components/Footer/Footer';
+// import Navigation from './components/Navbar/Navbar';
+// import Home from './pages/Home';
+// import Login from './components/LoginForm/LoginForm';
+// import Search from './pages/SearchPlantForm';
+// import Profile from './components/Profile/ProfilePage';
+// import './App.css';
 
+
+// const client = new ApolloClient({
+//   uri: '/graphql',
+//   cache: new InMemoryCache(),
+// });
+
+// function App() {
 //   return (
-//     <Router>
-//       <Switch>
-//         <Route path="/login">
-//           <Login />
-//         </Route>
-//         <Route path="/signup">
-//           <Signup />
-//         </Route>
-//         <Route path="/home">
-//           {isLoggedIn ? <Home /> : <Redirect to="/login" />}
-//         </Route>
-//         <Route path="/profile">
-//           {isLoggedIn ? <Profile /> : <Redirect to="/login" />}
-//         </Route>
-//         {/* Add more routes as needed */}
-//         <Route path="/">
-//           <Redirect to="/home" />
-//         </Route>
-//       </Switch>
-//     </Router>
+//     <div className="App">
+//       <ApolloProvider client={client}>
+//         <Router>
+//           <Header />
+//           <Navigation />
+//           <Routes>
+//             <Route path="/" element={<Home />} />
+//             <Route path="/login" element={<Login />} />
+//             <Route path="/search" element={<Search />} />
+//             <Route path="/profile" element={<Profile />} />
+//           </Routes>
+//           <Footer />
+//         </Router>
+//       </ApolloProvider>
+//     </div>
 //   );
 // }
 
 // export default App;
-///change to allow for apollo client, adjust other folders to allow for apollo as well 
-
-// import { ApolloProvider } from '@apollo/client';
+//edit comments
 import React from 'react';
-import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-import { ApolloProvider, ApolloClient, InMemoryCache } from '@apollo/client';
-import Header from './components/Header/Header';
-import Footer from './components/Footer/Footer';
-import Navigation from './components/Navbar/Navbar';
-import Home from './pages/Home';
-import Login from './components/LoginForm/LoginForm';
-import Search from './pages/SearchPlantForm';
-import Profile from './components/Profile/ProfilePage';
-import './App.css';
+import {
+  BrowserRouter as Router,
+  Routes,
+  Route,
+  
+} from 'react-router-dom';
 
+import Nav from './components/Nav';
+import Login from './pages/Login';
+import ImageFinder from './pages/ImageFinder';
+import Signup from './pages/Signup';
+import Plant from './pages/Plant'
+import Home from './pages/Home'
+
+import {
+  ApolloProvider,
+  ApolloClient,
+  InMemoryCache,
+  createHttpLink,
+} from '@apollo/client';
+import { setContext } from '@apollo/client/link/context';
+
+const httpLink = createHttpLink({
+  uri: '/graphql',
+});
+
+const authLink = setContext((_, { headers }) => {
+  const token = localStorage.getItem('id_token');
+  return {
+    headers: {
+      ...headers,
+      authorization: token ? `Bearer ${token}` : '',
+    },
+  };
+});
 
 const client = new ApolloClient({
-  uri: '/graphql',
+  link: authLink.concat(httpLink),
   cache: new InMemoryCache(),
 });
 
 function App() {
+  
   return (
-    <div className="App">
-      <ApolloProvider client={client}>
-        <Router>
-          <Header />
-          <Navigation />
-          <Routes>
-            <Route path="/" element={<Home />} />
-            <Route path="/login" element={<Login />} />
-            <Route path="/search" element={<Search />} />
-            <Route path="/profile" element={<Profile />} />
-          </Routes>
-          <Footer />
-        </Router>
-      </ApolloProvider>
-    </div>
+    <ApolloProvider client={client}>
+      <Router>
+        <Nav/>
+        <Routes>
+          <Route path='/' element={<Home />} />
+          <Route path='/ImageFinder/:username' element={<ImageFinder />} />
+          <Route path='/ImageFinder' element={<ImageFinder />} />
+          <Route path='/signup' element={<Signup />} />
+          <Route path='/Login' element={<Login />} />
+          <Route path='/plant/:id' element={<Plant />} />
+        </Routes>
+      </Router>
+    </ApolloProvider>
   );
 }
 
 export default App;
-//edit comments
