--- client/src/components/LoginForm/LoginForm.jsx
@@ -1,196 +1,0 @@
-import { useState } from 'react';
-import { Link } from 'react-router-dom';
-import { useMutation } from '@apollo/client';
-import { LOGIN_USER } from '../../utils/mutations';
-
-import Auth from '../../utils/auth';
-
-const Login = (props) => {
-  const [formState, setFormState] = useState({ email: '', password: '' });
-  const [login, { error, data }] = useMutation(LOGIN_USER);
-
-  // update state based on form input changes
-  const handleChange = (event) => {
-    const { name, value } = event.target;
-
-    setFormState({
-      ...formState,
-      [name]: value,
-    });
-  };
-
-  // submit form
-  const handleFormSubmit = async (event) => {
-    event.preventDefault();
-    console.log(formState);
-    try {
-      const { data } = await login({
-        variables: { ...formState },
-      });
-
-      Auth.login(data.login.token);
-    } catch (e) {
-      console.error(e);
-    }
-
-    // clear form values
-    setFormState({
-      email: '',
-      password: '',
-    });
-  };
-
-  return (
-    <main className="flex-row justify-center mb-4">
-      <div className="col-12 col-lg-10">
-        <div className="card">
-          <h4 className="card-header bg-dark text-light p-2">Login</h4>
-          <div className="card-body">
-            {data ? (
-              <p>
-                Success! You may now head{' '}
-                <Link to="/">back to the homepage.</Link>
-              </p>
-            ) : (
-              <form onSubmit={handleFormSubmit}>
-                <input
-                  className="form-input"
-                  placeholder="Your email"
-                  name="email"
-                  type="email"
-                  value={formState.email}
-                  onChange={handleChange}
-                />
-                <input
-                  className="form-input"
-                  placeholder="******"
-                  name="password"
-                  type="password"
-                  value={formState.password}
-                  onChange={handleChange}
-                />
-                <button
-                  className="btn btn-block btn-primary"
-                  style={{ cursor: 'pointer' }}
-                  type="submit"
-                >
-                  Submit
-                </button>
-              </form>
-            )}
-
-            {error && (
-              <div className="my-3 p-3 bg-danger text-white">
-                {error.message}
-              </div>
-            )}
-          </div>
-        </div>
-      </div>
-    </main>
-  );
-};
-
-export default Login;
-
-
-//fix comments, fix to be able to log in without error
-
-// I used my contact me from my portfolio to create this if it 
-//doesnt work then i need to adjust this page
-//see app and main
-// see SignupForm.js for comments
-
-
-// THIS IS A GUIDELINE FOR CHANGES
-// import { useState } from 'react';
-// import { Form, Button, Alert } from 'react-bootstrap';
-
-// import { loginUser } from '../utils/API';
-// import Auth from '../utils/auth';
-
-// const LoginForm = () => {
-//   const [userFormData, setUserFormData] = useState({ email: '', password: '' });
-//   const [validated] = useState(false);
-//   const [showAlert, setShowAlert] = useState(false);
-
-//   const handleInputChange = (event) => {
-//     const { name, value } = event.target;
-//     setUserFormData({ ...userFormData, [name]: value });
-//   };
-
-//   const handleFormSubmit = async (event) => {
-//     event.preventDefault();
-
-//     // check if form has everything (as per react-bootstrap docs)
-//     const form = event.currentTarget;
-//     if (form.checkValidity() === false) {
-//       event.preventDefault();
-//       event.stopPropagation();
-//     }
-
-//     try {
-//       const response = await loginUser(userFormData);
-
-//       if (!response.ok) {
-//         throw new Error('something went wrong!');
-//       }
-
-//       const { token, user } = await response.json();
-//       console.log(user);
-//       Auth.login(token);
-//     } catch (err) {
-//       console.error(err);
-//       setShowAlert(true);
-//     }
-
-//     setUserFormData({
-//       username: '',
-//       email: '',
-//       password: '',
-//     });
-//   };
-
-//   return (
-//     <>
-//       <Form noValidate validated={validated} onSubmit={handleFormSubmit}>
-//         <Alert dismissible onClose={() => setShowAlert(false)} show={showAlert} variant='danger'>
-//           Something went wrong with your login credentials!
-//         </Alert>
-//         <Form.Group className='mb-3'>
-//           <Form.Label htmlFor='email'>Email</Form.Label>
-//           <Form.Control
-//             type='text'
-//             placeholder='Your email'
-//             name='email'
-//             onChange={handleInputChange}
-//             value={userFormData.email}
-//             required
-//           />
-//           <Form.Control.Feedback type='invalid'>Email is required!</Form.Control.Feedback>
-//         </Form.Group>
-
-//         <Form.Group className='mb-3'>
-//           <Form.Label htmlFor='password'>Password</Form.Label>
-//           <Form.Control
-//             type='password'
-//             placeholder='Your password'
-//             name='password'
-//             onChange={handleInputChange}
-//             value={userFormData.password}
-//             required
-//           />
-//           <Form.Control.Feedback type='invalid'>Password is required!</Form.Control.Feedback>
-//         </Form.Group>
-//         <Button
-//           disabled={!(userFormData.email && userFormData.password)}
-//           type='submit'
-//           variant='success'>
-//           Submit
-//         </Button>
-//       </Form>
-//     </>
-//   );
-// };
-
-// export default LoginForm;
