--- client/src/components/Navbar/Navbar.css
@@ -1,4 +1,0 @@
-li {
-    color: white;
-    font-size: 30px;
-}
