+++ client/src/components/PlantCard/index.jsx
