+++ client/src/main.jsx
@@ -1,42 +1,21 @@
-// import React from 'react';
-// import ReactDOM from 'react-dom';
-// import App from './App';
-// import './index.css';
+// import ReactDOM from 'react-dom/client'
+// import App from './App.jsx'
+// import './index.css'
 
 // ReactDOM.createRoot(document.getElementById('root')).render(
-//   <React.StrictMode>
-//     <App />
-//   </React.StrictMode>
-// );
-// ////// move app.jsx here
-
-// import React from 'react';
-// import ReactDOM from 'react-dom';
-// import { BrowserRouter as Router } from 'react-router-dom';
-// import { ApolloProvider, ApolloClient, InMemoryCache } from '@apollo/client';
-
-// import App from './App.jsx';
-// import './index.css';
-
-// const client = new ApolloClient({
-//   uri: '/graphql',
-//   cache: new InMemoryCache(),
-// });
-
-// ReactDOM.createRoot(document.getElementById('root')).render(
-//   <React.StrictMode>
-//     <ApolloProvider client={client}>
-//       <Router>
-//         <App />
-//       </Router>
-//     </ApolloProvider>
-//   </React.StrictMode>
-// );
-import ReactDOM from 'react-dom/client'
-import App from './App.jsx'
-import './index.css'
-
-ReactDOM.createRoot(document.getElementById('root')).render(
-  <App />
-)
+//   <App />
+// )
 //edit comments
+import React from 'react';
+import ReactDOM from 'react-dom';
+import './index.css';
+import App from './App';
+import * as serviceWorker from './serviceWorker';
+
+ReactDOM.render(
+  <React.StrictMode>
+    <App />
+  </React.StrictMode>,
+  document.getElementById('root')
+);
+serviceWorker.register();
