+++ server/server.js
@@ -1,8 +1,6 @@
 // const express = require('express');
-// const { ApolloServer } = require('@apollo/server');
-// const { expressMiddleware } = require('@apollo/server/express4');
+// const { ApolloServer } = require('apollo-server-express');
 // const path = require('path');
-// const { authMiddleware } = require('./utils/auth');
 
 // const { typeDefs, resolvers } = require('./schemas');
 // const db = require('./config/connection');
@@ -43,48 +41,47 @@
 
 // // Call the async function to start the server
 //   startApolloServer();
-
-  // server/server.js
 const express = require('express');
 const { ApolloServer } = require('apollo-server-express');
 const path = require('path');
 
 const { typeDefs, resolvers } = require('./schemas');
+const { authMiddleware } = require('./utils/auth');
 const db = require('./config/connection');
-const app = express();
+
 const PORT = process.env.PORT || 3001;
+const app = express();
 
 const server = new ApolloServer({
   typeDefs,
   resolvers,
+  context: authMiddleware
 });
 
-// Create a new instance of an Apollo server with the GraphQL schema
-const startApolloServer = async () => {
+app.use(express.urlencoded({ extended: false }));
+app.use(express.json());
+
+app.use('/images', express.static(path.join(__dirname, '../client/images')));
+
+if (process.env.NODE_ENV === 'production') {
+  app.use(express.static(path.join(__dirname, '../client/build')));
+}
+
+app.get('*', (req, res) => {
+  res.sendFile(path.join(__dirname, '../client/build/index.html'));
+});
+
+const startApolloServer = async (typeDefs, resolvers) => {
   await server.start();
+  
+  server.applyMiddleware({ app });
 
-  app.use(express.urlencoded({ extended: false }));
-  app.use(express.json());
-
-  app.use('/graphql', expressMiddleware(server, {
-    context: authMiddleware
-  }));
-
-  if (process.env.NODE_ENV === 'production') {
-    app.use(express.static(path.join(__dirname, '../client/dist')));
-
-    app.get('*', (req, res) => {
-      res.sendFile(path.join(__dirname, '../client/dist/index.html'));
-    });
-  }
-
-  db.once('open', () => {
-    app.listen(PORT, () => {
-      console.log(`API server running on port ${PORT}!`);
-      console.log(`Use GraphQL at http://localhost:${PORT}/graphql`);
-    });
+db.once('open', () => {
+  app.listen(PORT, () => {
+    console.log(`API server running on port ${PORT}!`);
+    console.log(`Use GraphQL at http://localhost:${PORT}${server.graphqlPath}`);
   });
+})
 };
 
-// Call the async function to start the server
-  startApolloServer();
+startApolloServer(typeDefs, resolvers);
